requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
trl>=0.7.0
peft>=0.4.0
accelerate>=0.20.0

# Mathematical reasoning and evaluation
sympy>=1.12
numpy>=1.24.0
scipy>=1.10.0

# Logging and monitoring
wandb>=0.15.0
tensorboard>=2.13.0

# Data processing
pandas>=2.0.0
jsonlines>=3.1.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
click>=8.1.0
